Extract CanvasUserAvatar from CanvasPage
